requirements.txt

fastapi
uvicorn
openai-whisper
